Refactor base importer
